requirements.txt
torch>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
